# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards API
  description: Система управления банковскими картами
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  "/api/transactions/transfer":
    post:
      tags:
        - Транзакции
      summary: Перевод средств
      description: Выполняет перевод средств между картами
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TransactionRequest"
        required: true
      responses:
        '200':
          description: Перевод выполнен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TransactionDto"
        '400':
          description: Недостаточно средств
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TransactionDto"
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TransactionDto"
      security:
        - bearerAuth: []
  "/api/transactions/deposit":
    post:
      tags:
        - Транзакции
      summary: Пополнение счета (только ADMIN)
      description: Выполняет пополнение счета указанной карты
      operationId: createDeposit
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DepositRequest"
        required: true
      responses:
        '200':
          description: Пополнение выполнено
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TransactionDto"
        '400':
          description: Карта не найдена
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TransactionDto"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/TransactionDto"
      security:
        - bearerAuth: []
...
  "/api/transactions/user/{userId}":
    get:
      tags:
        - Транзакции
      summary: Получить транзакции по пользователю
      description: Возвращает список транзакций для указанного пользователя
      operationId: getByUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: date
        - name: direction
          in: query
          required: false
          schema:
            type: string
            default: desc
      responses:
        '200':
          description: Список транзакций пользователя
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelTransactionDto"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelTransactionDto"
      security:
        - bearerAuth: []
  "/api/transactions/own":
    get:
      tags:
        - Транзакции
      summary: Получить собственные транзакции
      description: Возвращает список транзакций текущего пользователя с пагинацией
      operationId: own
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: date
        - name: direction
          in: query
          required: false
          schema:
            type: string
            default: desc
      responses:
        '200':
          description: Список транзакций пользователя
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelTransactionDto"
      security:
        - bearerAuth: []
  "/api/transactions/card/{cardId}":
    get:
      tags:
        - Транзакции
      summary: Получить транзакции по карте
      description: Возвращает список транзакций для указанной карты
      operationId: getByCard
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: date
        - name: direction
          in: query
          required: false
          schema:
            type: string
            default: desc
      responses:
        '200':
          description: Список транзакций карты
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelTransactionDto"
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedModelTransactionDto"
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    TransactionDto:
      type: object
      description: DTO для представления транзакции
      properties:
        id:
          type: integer
          format: int64
          description: ID транзакции
        from:
          type: integer
          format: int64
          description: ID карты отправителя
        to:
          type: integer
          format: int64
          description: ID карты получателя
        amount:
          type: number
          description: Сумма транзакции
        comment:
          type: string
          description: Комментарий к транзакции
        date:
          type: string
          format: date-time
          description: Дата и время транзакции
    TransactionRequest:
      type: object
      description: Запрос на перевод средств
      properties:
        from:
          type: integer
          format: int64
          description: ID карты отправителя
          example: 1
        to:
          type: integer
          format: int64
          description: ID карты получателя
          example: 2
        amount:
          type: number
          description: Сумма перевода
          example: 500
        comment:
          type: string
          description: Комментарий к переводу
          example: Оплата услуг
          maxLength: 100
          minLength: 0
      required:
        - amount
        - comment
        - from
        - to
    DepositRequest:
      type: object
      description: Запрос на пополнение счета
      properties:
        card:
          type: integer
          format: int64
          description: ID карты для пополнения
          example: 1
        amount:
          type: number
          description: Сумма пополнения
          example: 1000
      required:
        - amount
        - card
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
...
    PagedModelTransactionDto:
      type: object
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/TransactionDto"
        page:
          "$ref": "#/components/schemas/PageMetadata"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT